requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
requests==2.31.0
aiogram==2.25.1